moves Measurement class validation tests out into their own file
